--- a/internal/storage/postgres/migrations/20231230174418_schema_version_index.sql
+++ b/internal/storage/postgres/migrations/20231230174418_schema_version_index.sql
@@ -1,0 +1,9 @@
+-- +goose Up
+-- +goose StatementBegin
+CREATE INDEX IF NOT EXISTS idx_schema_version ON schema_definitions (version);
+-- +goose StatementEnd
+
+-- +goose Down
+-- +goose StatementBegin
+DROP INDEX IF EXISTS idx_schema_version;
+-- +goose StatementEnd