--- a/internal/storage/postgres/migrations/20231230174418_schema_version_index.sql
+++ b/internal/storage/postgres/migrations/20231230174418_schema_version_index.sql
@@ -1,9 +1,10 @@
+-- +goose NO TRANSACTION
 -- +goose Up
 -- +goose StatementBegin
-CREATE INDEX IF NOT EXISTS idx_schema_version ON schema_definitions (version);
+CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_schema_version ON schema_definitions (version);
 -- +goose StatementEnd
 
 -- +goose Down
 -- +goose StatementBegin
-DROP INDEX IF EXISTS idx_schema_version;
+DROP INDEX CONCURRENTLY IF EXISTS idx_schema_version;
 -- +goose StatementEnd