@mohanish2504 2024-01-02
@tolgaOzen done

Test:
<code>
2024/01/02 18:59:12 OK   20230117143902_initial.sql (18.88ms)
2024/01/02 18:59:12 OK   20230117145140_multi_tenant.sql (11.81ms)
2024/01/02 18:59:12 OK   20230118185622_index_optimization.sql (10.94ms)
2024/01/02 18:59:12 OK   20230729145034_attributes.sql (7.92ms)
2024/01/02 18:59:12 OK   20231122120124_bundles.sql (5.03ms)
2024/01/02 18:59:12 OK   20231230174418_schema_version_index.sql (3.76ms)
2024/01/02 18:59:12 goose: successfully migrated database to version: 20231230174418
</code>